Simplified installation and uninstallation scripts, avoiding if-statements.

+++ install.sh
@@ -2,17 +2,14 @@
  
 # Find all dot files, back them up if they exist and create a symlink to the new dot file
 for file in $(find . -maxdepth 1 -name ".*" -type f -printf "%f\n" ); do
-	if [ -e ~/$file ]; then
-		mv -f ~/$file{,.bck}
-	fi
+	[ -e ~/$file ] &&	mv -f ~/$file{,.bck};
 	ln -s $PWD/$file ~/$file
 done
 
 # For all folders in .config, back them up if they exist and create a symlink to the new folder
+[ -d ~/.config ] || mkdir ~/.config;
 for folder in $(ls .config/ ); do
-	if [ -d ~/.config/$folder ]; then
-		mv -f ~/.config/$folder{,.bck}
-	fi
+	[ -d ~/.config/$folder ] && mv -f ~/.config/$folder{,.bck};
 	mkdir ~/.config/$folder
 	for file in $(ls .config/$folder ); do
 		ln -s $PWD/.config/$folder/$file ~/.config/$folder/$file
