+++ uninstall.sh
@@ -2,24 +2,16 @@
  
 # Find all dot files, delete them and restore the backup, if it exists
 for file in $(find . -maxdepth 1 -name ".*" -type f -printf "%f\n" ); do
-	if [ -h ~/$file ]; then
-		rm -f ~/$file
-	fi
-	if [ -e ~/${file}.bck ]; then
-		mv -f ~/$file{.bck,}
-	fi
+	[ -h ~/$file ] && rm -f ~/$file
+	[ -e ~/${file}.bck ] && mv -f ~/$file{.bck,};
 done
 
 # For all folders in .config, back them up if they exist and create a symlink to the new folder
 for folder in $(ls .config/ ); do
-	if [ -d ~/.config/$folder ]; then
-		rm -rdf ~/.config/$folder
-	fi
-	if [ -d ~/.config/${folder}.bck  ]; then
-		mv -f ~/.config/$folder{.bck,}
-	fi
+	[ -d ~/.config/$folder ] && rm -rdf ~/.config/$folder;
+	[ -d ~/.config/${folder}.bck  ] && mv -f ~/.config/$folder{.bck,};
 done
 
-rm -rd ~/.tmux
+rm -frd ~/.tmux
 
 echo "Uninstalled"
